Fix, update and consolidate sequence tests
